Tie drawer to search bar menu and make it a bottom sheet
