+++ android-app/app/src/main/res/layout/layout_home_bottom_sheet.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/homeBottomSheet"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+</FrameLayout>
